Add nameCell to shared tableUtils file
We will use this in the next commit
