+++ server/utils/tasks/table.test.ts
@@ -7,7 +7,6 @@ import {
   daysUntilDue,
   daysUntilDueCell,
   formatDaysUntilDueWithWarning,
-  nameCell,
   statusBadge,
   statusCell,
   taskTypeCell,
@@ -146,21 +145,6 @@ describe('table', () => {
     })
   })
 
-  describe('nameCell', () => {
-    it('returns the name of the person the task is assigned to as a TableCell object', () => {
-      const task = taskFactory.build()
-      expect(nameCell(task)).toEqual({ text: task.person.name })
-    })
-
-    it('returns an empty string as a TableCell object if the task doesnt have a person', () => {
-      const taskWithNoPersonName = taskFactory.build({ person: { name: undefined } })
-      const taskWithNoPerson = taskFactory.build({ person: undefined })
-
-      expect(nameCell(taskWithNoPersonName)).toEqual({ text: '' })
-      expect(nameCell(taskWithNoPerson)).toEqual({ text: '' })
-    })
-  })
-
   describe('daysUntilDueCell', () => {
     it('returns the days until due formatted for the UI as a TableCell object', () => {
       const task = taskFactory.build()
@@ -269,6 +253,7 @@ describe('table', () => {
         html: `<a href="/applications/${task.applicationId}/allocation" data-cy-taskId="${task.applicationId}">Allocate <span class="govuk-visually-hidden">task for ${task.person.name}</span></a>`,
       })
     })
+
     it('returns the cell when there is not a person present in the task', () => {
       const task = taskFactory.build({ person: undefined })
       expect(allocationLinkCell(task, 'Allocate')).toEqual({
